fix IPv6-only operation of Windows binaries
the SOCKET type is unsigned on Windows, and should be casted to an int before comparing with the highestfd variable (note: ``int highestfd = -1;'')

From /dev/humancontroller.

+++ code/qcommon/net_ip.c
@@ -1672,7 +1672,7 @@ void NET_Sleep(int msec)
 	{
 		FD_SET(ip6_socket, &fdr);
 		
-		if(ip6_socket > highestfd)
+		if((int)ip6_socket > highestfd)
 			highestfd = ip6_socket;
 	}
